+++ Backend/tienda/src/main/java/com/greenbuddies/store/StoreApplication.java
@@ -1,0 +1,29 @@
+package com.greenbuddies.store;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@SpringBootApplication
+public class StoreApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(StoreApplication.class, args);
+	}
+
+
+	@Bean
+	public WebMvcConfigurer corsConfigurer() {
+		return new WebMvcConfigurer() {
+			@Override
+			public void addCorsMappings(CorsRegistry registry) {
+				registry.addMapping("/*").allowedOrigins("").allowedMethods("*").allowedHeaders(
+						"*");
+			}
+		};
+	}
+
+
+}
