+++ Backend/tienda/src/main/java/com/greenbuddies/store/config/WebMvcConfiguration.java
@@ -1,0 +1,15 @@
+package com.greenbuddies.store.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.*;
+
+@Configuration
+@EnableWebMvc
+public class WebMvcConfiguration implements WebMvcConfigurer {
+
+    @Override
+    public void addViewControllers(final ViewControllerRegistry registry){
+        registry.addRedirectViewController("/swagger-ui.html", "/swagger-ui/index.html");
+    }
+
+}
