+++ Backend/tienda/src/main/java/com/greenbuddies/store/controller/ArticleController.java
@@ -1,0 +1,4 @@
+package com.greenbuddies.store.controller;
+
+public class ArticleController {
+}
