+++ Backend/tienda/src/main/java/com/greenbuddies/store/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.greenbuddies.store.controller;
+
+public class UserController {
+}
