+++ Backend/tienda/src/main/java/com/greenbuddies/store/exceptions/BadRequestException.java
@@ -1,0 +1,8 @@
+package com.greenbuddies.store.exceptions;
+
+
+public class BadRequestException extends Exception {
+    public BadRequestException(String mensaje){
+        super(mensaje);
+    }
+}
