+++ Backend/tienda/src/main/java/com/greenbuddies/store/repository/IProductRepository.java
@@ -1,0 +1,17 @@
+package com.greenbuddies.store.repository;
+
+import com.greenbuddies.store.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface IProductRepository extends JpaRepository<Product, Long> {
+    @Query("SELECT p FROM Product p WHERE p.name = ?1")
+    Optional<Product> findProductByName(String name);
+
+    @Query("SELECT p FROM Product p WHERE p.category.name = ?1")
+    List<Product> listProductsByCategory(String name);
+
+}
