+++ Backend/tienda/src/main/java/com/greenbuddies/store/repository/IUserRepository.java
@@ -1,0 +1,7 @@
+package com.greenbuddies.store.repository;
+
+import com.greenbuddies.store.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface IUserRepository extends JpaRepository<User, Long> {
+}
