--- Backend/tienda/src/main/java/com/greenbuddies/tienda/TiendaApplication.java
@@ -1,29 +1,0 @@
-package com.greenbuddies.tienda;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@SpringBootApplication
-public class TiendaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TiendaApplication.class, args);
-	}
-
-
-	@Bean
-	public WebMvcConfigurer corsConfigurer() {
-		return new WebMvcConfigurer() {
-			@Override
-			public void addCorsMappings(CorsRegistry registry) {
-				registry.addMapping("/*").allowedOrigins("").allowedMethods("*").allowedHeaders(
-						"*");
-			}
-		};
-	}
-
-
-}
