--- Backend/tienda/src/main/java/com/greenbuddies/tienda/config/SpringConfig.java
@@ -1,23 +1,0 @@
-package com.greenbuddies.tienda.config;
-
-import org.modelmapper.ModelMapper;
-import org.modelmapper.config.Configuration;
-import org.modelmapper.convention.NamingConventions;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-
-@SpringBootApplication
-
-public class SpringConfig {
-
-    @Bean
-    public ModelMapper getModelMapper(){
-        ModelMapper modelMapper = new ModelMapper();
-        modelMapper.getConfiguration()
-                .setFieldMatchingEnabled(true)
-                .setFieldAccessLevel(Configuration.AccessLevel.PRIVATE)
-                .setSourceNamingConvention(NamingConventions.JAVABEANS_MUTATOR);
-        return modelMapper;
-    }
-}
