--- Backend/tienda/src/main/java/com/greenbuddies/tienda/config/SwaggerConfig.java
@@ -1,24 +1,0 @@
-package com.greenbuddies.tienda.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(
-                        RequestHandlerSelectors
-                                .basePackage("com.greenbuddies.tienda"))
-                .paths(PathSelectors.any())
-                .build();
-    }
-}
