--- Backend/tienda/src/main/java/com/greenbuddies/tienda/config/WebMvcConfiguration.java
@@ -1,15 +1,0 @@
-package com.greenbuddies.tienda.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.*;
-
-@Configuration
-@EnableWebMvc
-public class WebMvcConfiguration implements WebMvcConfigurer {
-
-    @Override
-    public void addViewControllers(final ViewControllerRegistry registry){
-        registry.addRedirectViewController("/swagger-ui.html", "/swagger-ui/index.html");
-    }
-
-}
