--- Backend/tienda/src/main/java/com/greenbuddies/tienda/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.greenbuddies.tienda.controller;
-
-public class UserController {
-}
