--- Backend/tienda/src/main/java/com/greenbuddies/tienda/exceptions/BadRequestException.java
@@ -1,8 +1,0 @@
-package com.greenbuddies.tienda.exceptions;
-
-
-public class BadRequestException extends Exception {
-    public BadRequestException(String mensaje){
-        super(mensaje);
-    }
-}
