--- Backend/tienda/src/main/java/com/greenbuddies/tienda/model/Article.java
@@ -1,47 +1,0 @@
-package com.greenbuddies.tienda.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.*;
-import org.hibernate.Hibernate;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "blog_article")
-public class Article {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name= "blog_article_id")
-    private Long id;
-
-    private String title;
-    private String content;
-
-    @OneToMany(cascade=CascadeType.ALL)
-    @JoinColumn(name="blog_article_id")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-    @ToString.Exclude
-    private List<Image> images = new ArrayList<>();
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Article article = (Article) o;
-        return id != null && Objects.equals(id, article.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
