--- Backend/tienda/src/main/java/com/greenbuddies/tienda/model/Cart.java
@@ -1,49 +1,0 @@
-package com.greenbuddies.tienda.model;
-
-
-import lombok.*;
-import org.hibernate.Hibernate;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-@Entity
-@Table(name = "cart")
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-public class Cart {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name= "idCart")
-    private Long id;
-
-    @OneToOne(mappedBy = "cart")
-    private User user;
-
-    @OneToOne
-    @JoinColumn(name = "factura_id", nullable = false)
-    private Receipt receipt;
-
-
-    @ManyToMany(mappedBy = "carts")
-    @ToString.Exclude
-    private List<Product> products = new ArrayList<>();
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Cart cart = (Cart) o;
-        return id != null && Objects.equals(id, cart.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
