--- Backend/tienda/src/main/java/com/greenbuddies/tienda/model/Category.java
@@ -1,47 +1,0 @@
-package com.greenbuddies.tienda.model;
-
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.*;
-import org.hibernate.Hibernate;
-
-import javax.persistence.*;
-import java.util.List;
-import java.util.Objects;
-
-@Entity
-@Table(name = "category")
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-public class Category {
-
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name= "idCategory")
-    private Long id;
-
-    private String name;
-    private String description;
-
-    @OneToMany(mappedBy = "category")//orphanRemoval = true
-    @JsonIgnore
-    @ToString.Exclude
-    private List<Product> products;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Category category = (Category) o;
-        return id != null && Objects.equals(id, category.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
