--- Backend/tienda/src/main/java/com/greenbuddies/tienda/model/Product.java
@@ -1,62 +1,0 @@
-package com.greenbuddies.tienda.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.*;
-import org.hibernate.Hibernate;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-@Entity
-@Table(name = "product")
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name= "idProduct")
-    private Long id;
-    private String name;
-    private Double price;
-    private int quantity;
-    private String description;
-
-
-    @ManyToOne//(fetch = FetchType.LAZY)
-    @JoinColumn(name = "category_id", nullable = false)
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-    private Category category;
-
-    @OneToMany(cascade=CascadeType.ALL)
-    @JoinColumn(name="product_id")
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-    @ToString.Exclude
-    private List<Image> images = new ArrayList<>();
-
-    @JoinTable(name = "products_cart",
-            joinColumns = {@JoinColumn(name = "product_id")},
-            inverseJoinColumns = {@JoinColumn(name = "cart_id", nullable = false)}
-    )
-    @ManyToMany(cascade = CascadeType.MERGE)
-    @ToString.Exclude
-    private List<Cart> carts = new ArrayList<>();
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Product product = (Product) o;
-        return id != null && Objects.equals(id, product.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
