--- Backend/tienda/src/main/java/com/greenbuddies/tienda/model/Receipt.java
@@ -1,41 +1,0 @@
-package com.greenbuddies.tienda.model;
-
-import lombok.*;
-import org.hibernate.Hibernate;
-
-import javax.persistence.*;
-import java.util.Objects;
-
-
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "receipt")
-public class Receipt {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name= "idReceipt")
-    private Long id;
-
-    private Integer number;
-    private Double total;
-
-    @OneToOne(mappedBy = "receipt")
-    private Cart cart;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        Receipt receipt = (Receipt) o;
-        return id != null && Objects.equals(id, receipt.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
