--- Backend/tienda/src/main/java/com/greenbuddies/tienda/model/User.java
@@ -1,62 +1,0 @@
-package com.greenbuddies.tienda.model;
-
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.*;
-import org.hibernate.Hibernate;
-
-import javax.persistence.*;
-import java.util.Objects;
-
-@Getter
-@Setter
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "user")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name= "idUser")
-    private Long id;
-    private String name;
-    private String lastName;
-    private String email;
-    private String password;
-    private String birthDate;
-
-    @ManyToOne//(fetch = FetchType.LAZY)
-    @JoinColumn(name = "country_id", nullable = false)
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-    private Country country;
-
-    @ManyToOne//(fetch = FetchType.LAZY)
-    @JoinColumn(name = "state_id", nullable = false)
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-    private State state;
-
-    @ManyToOne//(fetch = FetchType.LAZY)
-    @JoinColumn(name = "zip_code_id", nullable = false)
-    @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
-    private ZipCode zipCode;
-
-    @OneToOne(cascade = CascadeType.ALL)
-    @JoinColumn(name = "cart_id", nullable = false)
-    private Cart cart;
-
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
-        User user = (User) o;
-        return id != null && Objects.equals(id, user.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return getClass().hashCode();
-    }
-}
