--- Backend/tienda/src/main/java/com/greenbuddies/tienda/repository/IProductRepository.java
@@ -1,17 +1,0 @@
-package com.greenbuddies.tienda.repository;
-
-import com.greenbuddies.tienda.model.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface IProductRepository extends JpaRepository<Product, Long> {
-    @Query("SELECT p FROM Product p WHERE p.name = ?1")
-    Optional<Product> findProductByName(String name);
-
-    @Query("SELECT p FROM Product p WHERE p.category.name = ?1")
-    List<Product> listProductsByCategory(String name);
-
-}
