--- Backend/tienda/src/main/java/com/greenbuddies/tienda/repository/IUserRepository.java
@@ -1,7 +1,0 @@
-package com.greenbuddies.tienda.repository;
-
-import com.greenbuddies.tienda.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface IUserRepository extends JpaRepository<User, Long> {
-}
