+++ Backend/tienda/src/test/java/com/greenbuddies/store/StoreApplicationTests.java
@@ -1,0 +1,13 @@
+package com.greenbuddies.store;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class StoreApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
