--- Backend/tienda/src/test/java/com/greenbuddies/tienda/TiendaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.greenbuddies.tienda;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TiendaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
